1.0
There is no goal, pickup ammo on spawn (wire-spheres) shoot bots. There is a section of the wall that shatters when hit,  acts as a gate. When launched the game starts, the game does not end, relaunch to restart.
Controls:
WASD - movement
Mouse XYZ - look
Mouse 1 and 2 - shoot light and heavy
E - lift (for cube)
F - flashlight
T - "grapple"
G - sticky grenade

This is all a work in progress